+++ wyoming_faster_whisper/__init__.py
@@ -1,4 +1,5 @@
 """Wyoming server for faster-whisper."""
+
 from importlib.metadata import version
 
 __version__ = version("wyoming_faster_whisper")