+++ wyoming_faster_whisper/const.py
@@ -1,0 +1,41 @@
+"""Constants."""
+
+from enum import Enum
+
+
+class SttLibrary(str, Enum):
+    """Speech-to-text library."""
+
+    AUTO = "auto"
+    FASTER_WHISPER = "faster-whisper"
+    TRANSFORMERS = "transformers"
+    SHERPA = "sherpa"
+
+
+PARAKEET_LANGUAGES = {
+    "bg",
+    "hr",
+    "cs",
+    "da",
+    "nl",
+    "en",
+    "et",
+    "fi",
+    "fr",
+    "de",
+    "el",
+    "hu",
+    "it",
+    "lv",
+    "lt",
+    "mt",
+    "pl",
+    "pt",
+    "ro",
+    "sk",
+    "sl",
+    "es",
+    "sv",
+    "ru",
+    "uk",
+}